move useFormattedDuration hook into its own file
