+++ src/components/RecordMoviesComponent/index.tsx
@@ -5,6 +5,7 @@ import RecordMoviesModal from "../RecordMoviesModal";
 import ViewportButton from "../ViewportButton";
 
 import styles from "./style.css";
+import { useFormattedDuration } from "./useFormattedDuration";
 
 interface RecordMovieComponentProps {
     movieUrl: string;
@@ -93,34 +94,34 @@ const RecordMovieComponent = (props: RecordMovieComponentProps) => {
         return "Start movie recording";
     };
 
-    /**
-     * onloadedmetadata is async hence the need for this custom hook
-     * to retrieve the duration from the recorded movie.
-     */
-    const useFormattedDuration = () => {
-        const [formattedDuration, setFormattedDuration] = useState("");
+    // /**
+    //  * onloadedmetadata is async hence the need for this custom hook
+    //  * to retrieve the duration from the recorded movie.
+    //  */
+    // export const useFormattedDuration = (movieUrl: string) => {
+    //     const [formattedDuration, setFormattedDuration] = useState("");
 
-        useEffect(() => {
-            const video = document.createElement("video");
-            video.src = movieUrl;
-            video.onloadedmetadata = () => {
-                const durationInSeconds = video.duration;
-                const minutes = Math.floor((durationInSeconds % 3600) / 60);
-                const secs = Math.ceil(durationInSeconds) % 60;
+    //     useEffect(() => {
+    //         const video = document.createElement("video");
+    //         video.src = movieUrl;
+    //         video.onloadedmetadata = () => {
+    //             const durationInSeconds = video.duration;
+    //             const minutes = Math.floor((durationInSeconds % 3600) / 60);
+    //             const secs = Math.ceil(durationInSeconds) % 60;
 
-                const minutesStr = minutes.toString().padStart(2, "0");
-                const secondsStr = secs.toString().padStart(2, "0");
+    //             const minutesStr = minutes.toString().padStart(2, "0");
+    //             const secondsStr = secs.toString().padStart(2, "0");
 
-                setFormattedDuration(`${minutesStr}:${secondsStr}`);
-                video.src = "";
-                video.load();
-            };
-        }, [movieUrl]);
+    //             setFormattedDuration(`${minutesStr}:${secondsStr}`);
+    //             video.src = "";
+    //             video.load();
+    //         };
+    //     }, [movieUrl]);
 
-        return formattedDuration;
-    };
+    //     return formattedDuration;
+    // };
 
-    const duration = useFormattedDuration();
+    const duration = useFormattedDuration(movieUrl);
 
     return (
         <div className={styles.record}>
